client-api: Split ErrorKind and tests to submodules
The `error` module is very big and `ErrorKind` and associated types are the parts
that take most of the place so they deserve their own module.
